+++ restapi/lib/RestApiPOSTMethods.php
@@ -1,0 +1,7 @@
+<?
+namespace Rest;
+require_once($_SERVER['DOCUMENT_ROOT'] . "/bitrix/modules/main/include/prolog_before.php");
+
+class RestApiPOSTMethods {
+	
+}
