+++ restapi/lib/srp.php
@@ -1,0 +1,11 @@
+<?namespace Rest;
+require_once($_SERVER['DOCUMENT_ROOT'] . "/bitrix/modules/main/include/prolog_before.php");
+
+class SRP {
+    public static function get(){
+        /*
+        *   Какой то код для проверки  SRP
+        */
+        return true;
+    }
+}
